+++ src/Classifiers/MultilayerPerceptron.php
@@ -138,7 +138,7 @@ class MultilayerPerceptron implements Estimator, Learner, Iterative, Online, Pro
     protected int $evalInterval;
 
     /**
-     * The number of epochs without improvement in the validation score to wait before considering an early stop.
+     * The number of evaluations without improvement in the validation score to wait before considering an early stop.
      *
      * @var positive-int
      */
@@ -229,9 +229,9 @@ class MultilayerPerceptron implements Estimator, Learner, Iterative, Online, Pro
         ?Optimizer $optimizer = null,
         ?float $maxGradientNorm = null,
         int $epochs = 1000,
-        float $minChange = 1e-4,
-        int $evalInterval = 3,
-        int $window = 5,
+        float $minChange = 1e-5,
+        int $evalInterval = 1,
+        int $window = 10,
         float $holdOut = 0.1,
         ?ClassificationLoss $costFn = null,
         ?Metric $metric = null,