+++ src/Classifiers/SoftmaxClassifier.php
@@ -117,7 +117,7 @@ class SoftmaxClassifier implements Estimator, Learner, Iterative, Online, Probab
     protected int $evalInterval;
 
     /**
-     * The number of epochs without improvement in the validation score to wait before considering an early stop.
+     * The number of evaluations without improvement in the validation score to wait before considering an early stop.
      *
      * @var positive-int
      */
@@ -199,9 +199,9 @@ class SoftmaxClassifier implements Estimator, Learner, Iterative, Online, Probab
         float $l1Penalty = 1e-4,
         float $l2Penalty = 1e-4,
         int $epochs = 1000,
-        float $minChange = 1e-4,
-        int $evalInterval = 3,
-        int $window = 5,
+        float $minChange = 1e-5,
+        int $evalInterval = 1,
+        int $window = 10,
         float $holdOut = 0.1,
         ?ClassificationLoss $costFn = null,
         ?Metric $metric = null